Move encounter components into their own directory
