+++ src/components/encounters/CreateEncounterForm.tsx
@@ -1,5 +1,5 @@
 import React, { FormEvent, FunctionComponentElement, useState } from 'react'
-import api from '../api'
+import api from '../../api'
 import Swal from 'sweetalert2'
 import {
   Card,
